feat: reuse saved cards for group funding
