--- .github/workflows/apply-saved-card-group-funding.yml
@@ -1,31 +1,0 @@
-name: Apply saved-card group funding patch
-
-on:
-  push:
-    branches:
-      - agent/reuse-saved-card-for-group-funding
-    paths:
-      - agent-scripts/apply-saved-card-group-funding.py
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: agent/reuse-saved-card-for-group-funding
-          fetch-depth: 0
-      - name: Apply patch
-        run: python agent-scripts/apply-saved-card-group-funding.py
-      - name: Commit implementation
-        run: |
-          rm agent-scripts/apply-saved-card-group-funding.py
-          rm .github/workflows/apply-saved-card-group-funding.yml
-          git config user.name "OpenAI"
-          git config user.email "openai@users.noreply.github.com"
-          git add -A
-          git commit -m "feat: reuse saved cards for group funding"
-          git push origin HEAD:agent/reuse-saved-card-for-group-funding
